# src/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PautaPro API",
    "description": "API Empresarial para Integración de Sistemas de Gestión Publicitaria",
    "version": "2.0.0",
    "contact": {
      "name": "API Support",
      "email": "api-support@pautapro.com",
      "url": "https://developers.pautapro.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.pautapro.com/v2",
      "description": "Servidor de Producción"
    },
    {
      "url": "https://api-staging.pautapro.com/v2",
      "description": "Servidor de Staging"
    },
    {
      "url": "http://localhost:3001/api/v2",
      "description": "Servidor de Desarrollo Local"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key para autenticación empresarial"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT para usuarios autenticados"
      }
    },
    "schemas": {
...
      "Orden": {
        "type": "object",
        "required": ["id_cliente", "numero_orden", "tipo_orden"],
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID único de la orden",
            "example": 67890
          },
          "id_cliente": {
            "type": "integer",
            "description": "ID del cliente",
            "example": 12345
          },
          "id_campania": {
            "type": "integer",
            "description": "ID de la campaña",
            "example": 987
          },
          "numero_orden": {
            "type": "string",
            "description": "Número de orden único",
            "example": "ORD-2024-001234"
          },
          "tipo_orden": {
            "type": "string",
            "description": "Tipo de orden",
            "example": "digital",
            "enum": ["digital", "tv", "radio", "print", "outdoor"]
          },
          "descripcion": {
            "type": "string",
            "description": "Descripción de la orden",
            "example": "Campaña de lanzamiento producto X"
          },
          "observaciones": {
            "type": "string",
            "description": "Observaciones adicionales",
            "example": "Cliente requiere aprobación previa"
          },
          "fecha_inicio": {
            "type": "string",
            "format": "date",
            "description": "Fecha de inicio",
            "example": "2024-11-15"
          },
          "fecha_termino": {
            "type": "string",
            "format": "date",
            "description": "Fecha de término",
            "example": "2024-12-15"
          },
          "total": {
            "type": "number",
            "format": "decimal",
            "description": "Total de la orden",
            "example": 1500000.00
          },
          "iva": {
            "type": "number",
            "format": "decimal",
            "description": "IVA de la orden",
            "example": 285000.00
          },
          "descuento": {
            "type": "number",
            "format": "decimal",
            "description": "Descuento aplicado",
            "example": 50000.00
          },
          "recargo": {
            "type": "number",
            "format": "decimal",
            "description": "Recargo aplicado",
            "example": 0.00
          },
          "estado": {
            "type": "string",
            "description": "Estado de la orden",
            "example": "pendiente",
            "enum": ["pendiente", "activa", "completada", "cancelada"]
          },
          "fecha_creacion": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de creación",
            "example": "2024-11-13T17:29:40.340Z"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/ordenes": {
      "get": {
        "summary": "Listar órdenes",
        "description": "Obtiene una lista paginada de órdenes con filtros avanzados",
        "tags": ["Órdenes"],
        "security": ["ApiKeyAuth"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Número de página",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Límite de elementos por página",
            "schema": {
              "type": "integer",
              "default": 50
            }
          },
          {
            "name": "estado",
            "in": "query",
            "description": "Filtrar por estado",
            "schema": {
              "type": "string",
              "enum": ["pendiente", "activa", "completada", "cancelada"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de órdenes obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Orden"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear nueva orden",
        "description": "Crea una nueva orden en el sistema",
        "tags": ["Órdenes"],
        "security": ["ApiKeyAuth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id_cliente", "numero_orden", "tipo_orden"],
                "properties": {
                  "id_cliente": {
                    "type": "integer"
                  },
                  "numero_orden": {
                    "type": "string"
                  },
                  "tipo_orden": {
                    "type": "string",
                    "enum": ["digital", "tv", "radio", "print", "outdoor"]
                  },
                  "descripcion": {
                    "type": "string"
                  },
                  "total": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Orden creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Orden"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}